StaggerVFX
Created a circle of broken music notes that will circle around an enemy's head while they are staggered.
staggerVFX Added to Bassics, Small Fry, King Sal, and the dummy enemy
Only Bassics has the VFX in the proper position because he is the only enemy that has a stagger animation. Remember to change the position of the VFX to where the head is during the stagger animation